Add definitions for resource attributes.
These are the attributes that can be applied to the resource file/fork
or the resource themselves. These are not yet bound to anything, or
affecting functionality. Not all of these are applicable on a system
that isn't Mac OS Classic, and others such as the compact flag are
implemented in an unknown way.